requirements.txt
streamlit
google-genai
Pillow
gTTS